requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-login==0.6.3
redis==5.2.1
python-dotenv==1.1.0
eventlet==0.39.1
